Created a DB config template set for MySQL.
The config/database.yml file was changed to use MySQL instead of
sqlite3 and renamed to config/database.yml.template. The database.yml
file is now in .gitignore so we don't accidentally commit MySQL
credentials into git. Also, the README file was altered to reflect the
DB configuration changes and the "Getting Started" section was
updated.