Only display public and own resources in list.
The All Resources view now only lists public resources and those that
the user owns. This implementation is extremely unoptimized. It should
be refactored soon.

This may not be a complete solution since if an unauthorized user has
the id of a resource, they may be able to view it. Needs more
investigation.